Better support for Electric Cloud's emake and Mac OS X
Electric Cloud's emake did not originally support $(eval), but that has
changed with recent versions. This version of GMSL detects emake and
checks the version number so that with recent versions of emake all
GMSL functions are available.

It was not possible to run the test suite on Mac OS X because the md5sum
utility does not exist there (it's called md5 instead). Now the test
suite detects whether md5 or md5sum is available.