Fix bug #12. set function no longer returns a value.
When the memoize function was added the set function was changed to return
a value. This causes breakages where set is used without consuming the
return value because GNU Make will attempt to parse the return value.

Restore set to the previous functionality where it did not return any value
and add a wrapper function __gmsl_set that does return a value (and calls
set) which is used by memoize.